backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.13.0
djoser==2.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
pybase64==1.3.2
django-filter==23.1
django-urlshortner